refactor: 구버전 Subscribe Controller 분리
